+++ src/main/java/org/example/statistics/StatisticsDAO.java
@@ -23,7 +23,7 @@ public class StatisticsDAO {
 
         try {
             // Fetch statistics from the users table
-            Result<Record6<Integer, BigDecimal, BigDecimal, BigDecimal, BigDecimal, BigDecimal>> result = dslContext.select(
+            Result<Record6<Integer, Double, Double, Double, Double, Double>> result = dslContext.select(
                             USERS.TOTAL_BETS,
                             USERS.TOTAL_STAKES,
                             USERS.TOTAL_WINNINGS,
@@ -35,14 +35,14 @@ public class StatisticsDAO {
                     .where(USERS.USER_ID.eq(userId))
                     .fetch();
 
-            for (Record6<Integer, BigDecimal, BigDecimal, BigDecimal, BigDecimal, BigDecimal> record : result) {
+            for (Record6<Integer, Double, Double, Double, Double, Double> record : result) {
                 Statistics stat = new Statistics();
                 stat.setTotalBets(record.get(USERS.TOTAL_BETS));  // TOTAL_BETS is an Integer
-                stat.setTotalStakes(record.get(USERS.TOTAL_STAKES).doubleValue());  // TOTAL_STAKES is a BigDecimal
-                stat.setTotalWinnings(record.get(USERS.TOTAL_WINNINGS).doubleValue());  // TOTAL_WINNINGS is a BigDecimal
-                stat.setWinRate(record.get(USERS.WIN_RATE).doubleValue());  // WIN_RATE is a BigDecimal
-                stat.setAverageBet(record.get(USERS.AVERAGE_BET).doubleValue());  // AVERAGE_BET is a BigDecimal
-                stat.setMaxBet(record.get(USERS.MAX_BET).doubleValue());  // MAX_BET is a BigDecimal
+                stat.setTotalStakes(record.get(USERS.TOTAL_STAKES));  // TOTAL_STAKES is a BigDecimal
+                stat.setTotalWinnings(record.get(USERS.TOTAL_WINNINGS));  // TOTAL_WINNINGS is a BigDecimal
+                stat.setWinRate(record.get(USERS.WIN_RATE));  // WIN_RATE is a BigDecimal
+                stat.setAverageBet(record.get(USERS.AVERAGE_BET));  // AVERAGE_BET is a BigDecimal
+                stat.setMaxBet(record.get(USERS.MAX_BET));  // MAX_BET is a BigDecimal
 
                 stats.add(stat);
             }
